PLEASE READ IF YOU EDIT ANY SCENE :
I modified the camera prefab in a way that's gonna be easier to apply
effects to it. The new prefab to use is under
StandardAssets/Camera/. Please replace the Camera in the scene you're
editing with it and Link it to the Player again. Thank you !

- Added Screen Shake effect on player being damaged and damaging other
  enemies.
- Added Freeze effect on player being damaged and damaing other enemies.

These functions can be found in CameraScript and GameManager respectively.
They're pretty simple and commented if you want to use them. Be carefull
though, these effects must be used in moderation !